add utils module, clean up imports
